requirements.txt
influxdb
ruuvitag-sensor
paho-mqtt
requests
